Add files via upload
These files contain all matrices of containment exponents. Each matrix is of size t_n x t_n where t_n is the number of full binary rooted trees with n leaves. For a fixed n we have n! such matrices, applying the permutation to the leaves of the second tree in the comparison.